+++ pkg/analyzer/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Deprecated `ResourceProviderMixin.newFile2`, use `newFile` instead.
 * Deprecated `ResourceProviderMixin.newAnalysisOptionsYamlFile2`, use `newAnalysisOptionsYamlFile` instead.
 * Deprecated `DartType.resolveToBound`, use `TypeSystem.resolveToBound` instead.
+* Deprecated `LibraryElement.getImportsWithPrefix`, use `PrefixElement.imports` instead.
 
 ## 4.0.0
 * Removed deprecated `UriKind` and `Source.uriKind`.